% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{Add plus 2 to a number}
\usage{
add_two(x)
}
\arguments{
\item{x}{A number (integer)}
}
\value{
a number object
}
\description{
This functoin adds 2 to any number provided.
}
\examples{
add_two(2)
add_two(98)
add_two(-8)
}
