% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_1.R
\docType{data}
\name{df_1}
\alias{df_1}
\title{truncated_mtcars}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{\code{mpg}}{double Miles/(US) gallon}
  \item{\code{cyl}}{double Number of cylindersN}
  \item{\code{disp}}{double COLUMN_DESCRIPTION}
  \item{\code{hp}}{double COLUMN_DESCRIPTION}
  \item{\code{drat}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://www.datarecode.com/}
}
\usage{
df_1
}
\description{
The data was extracted from the 1974 Motor Trend US magazine
}
\keyword{datasets}
